fix: structure
- update folders
- fix import modules
- testing local
