+++ node-js/src/api/routes.js
@@ -1,3 +1,13 @@
-const enrollToWaitingList = require('../controllers/enroll/enrollToWaitingList')
+import EnrollController from '../controllers/enroll/enroll-controller.js';
+import express from 'express';
+import rescue from 'express-rescue';
 
-module.exports = { enrollToWaitingList }
+const apiRoutes = express.Router();
+
+apiRoutes.get("/", (_req, res ) => {
+  res.sendStatus(200);
+});
+
+apiRoutes.post('/enroll', rescue(EnrollController.postEnroll));
+
+export default { apiRoutes }
