+++ node-js/src/service/enroll/enroll-service.js
@@ -1,0 +1,5 @@
+
+class EnrollService {
+};
+
+export default EnrollService;
